fix: store values as order-preserving 64-bit codes (artifact format 2.0)
Ordering operators were meaningless on three of the six column types. STRING,
DECIMAL and TIMESTAMP cells were stored as dictionary IDs assigned in first-seen
order, so GT/GTE/LT/LTE and the range operators compared insertion order rather
than value: "AMOUNT GT 100.00" did not match 1000.00, and "NAME GT MMM" matched
nothing at all, because an input the compiler had never seen resolved to the
null ID. DECIMAL equality was string identity, so 0.50 did not equal 0.5. And
INTEGER inputs were narrowed through Number.intValue(), so 2^32+1 matched a rule
written for 1.

Every column type now encodes into one order-preserving 64-bit domain, so
comparing two codes is comparing two values: strings by rank in the (now sorted)
dictionary, decimals as the unscaled value at a per-column scale, dates as epoch
days, timestamps as epoch microseconds. Codes are doubled, which leaves the odd
numbers to mean "strictly between two representable values" - that is how an
input the artifact cannot represent exactly (a string absent from the
dictionary, a decimal finer than the column's scale) still orders correctly
against every stored value while comparing equal to none of them.

Artifact format goes to 2.0 and is not readable by 1.x: value slots widen from
4 to 8 bytes, column definitions carry a decimal scale (12 -> 16 bytes), and
dictionary entries are written in sorted order.

Two consequences worth knowing about, both recorded in the review doc:

- Artifacts are ~1.8x larger, which brings the format's 2 GB ceiling closer.
  That ceiling was already a known limit; it is now the binding one at PRD
  scale.
- Dictionary heap roughly halved (24.1 MB -> 10.8 MB at 200K rows). Sorted IDs
  let the loader resolve a string by binary search over the entries it already
  holds, so the reverse HashMap is gone, and DECIMAL/TIMESTAMP values no longer
  enter the dictionary at all.

API changes that follow from storing these types numerically: DECIMAL outputs
decode as BigDecimal at the column's scale rather than String, INTEGER outputs
as Long, and DATE inputs must be a LocalDate - the old "bare Integer means epoch
day" path silently read a year as a day number.

Findings and reproductions are in docs/review/engine-design-review.md (2.2),
with the before/after table in section 8.

+++ kisoku-runtime/src/main/java/in/systemhalted/kisoku/runtime/loader/ColumnDecoder.java
@@ -28,8 +28,8 @@ sealed interface ColumnDecoder
   /**
    * Check if an already-coerced input value matches the condition at the given row.
    *
-   * <p>Equivalent to {@link #matches(int, Object)} but takes the comparable int produced by {@link
-   * TypeCoercion#toComparableInt} directly, skipping per-call coercion. The columnar bulk kernel
+   * <p>Equivalent to {@link #matches(int, Object)} but takes the comparable code produced by {@link
+   * TypeCoercion#toComparableCode} directly, skipping per-call coercion. The columnar bulk kernel
    * uses this to verify survivors from a batch of pre-coerced codes without re-coercing.
    *
    * <p>{@code present} must be carried alongside the code because the coerced domain cannot express
@@ -37,11 +37,11 @@ sealed interface ColumnDecoder
    * only the latter may satisfy a condition.
    *
    * @param rowIndex the row to check
-   * @param coercedValue the input value already coerced to its comparable int
+   * @param coercedValue the input value already coerced to its comparable code
    * @param present whether the input actually supplied a value for this column
    * @return true if matches, false otherwise
    */
-  boolean matchesCoerced(int rowIndex, int coercedValue, boolean present);
+  boolean matchesCoerced(int rowIndex, long coercedValue, boolean present);
 
   /**
    * Check if this row has a condition (not blank).
